EmoRecAI/requirements.txt
numpy>=1.23
pandas>=1.5
pillow>=9.0
opencv-python>=4.6
seaborn>=0.13
matplotlib>=3.5
scikit-learn>=1.1
tensorflow>=2.10,<3.0
keras>=2.11 